feature(combination): separate base combination class and groe combination
