+++ darts/models/groe_combination_model.py
@@ -1,48 +1,19 @@
 """
-Standard Regression model
+GROE combination model
 -------------------------
 """
 
 from ..timeseries import TimeSeries
-from ..logging import get_logger, raise_log, raise_if_not
+from ..logging import get_logger
 from ..models.forecasting_model import ForecastingModel
+from ..models.combination_model import CombinationModel
 from ..metrics import smape
 from typing import List, Callable
 import numpy as np
-import pandas as pd
 
 from ..utils.cross_validation import generalized_rolling_origin_evaluation as groe
 
 logger = get_logger(__name__)
-
-
-class CombinationModel:
-    """
-    Base class for combination models.
-    Acts as a constant weight combination model that averages all predictions from `models`
-    """
-    def __init__(self, models: List[ForecastingModel]):
-        for model in models:
-            raise_if_not(isinstance(model, ForecastingModel),
-                         "All models must be instances of forecasting models from darts.models")
-        self.models = models
-        self.weights = np.ones(len(self.models)) / len(self.models)
-        self.train_ts = None
-        self.predictions = None
-        self._fit_called = False
-
-    def fit(self, train_ts: TimeSeries) -> None:
-        self.train_ts = train_ts
-        self._fit_called = True
-
-    def predict(self, n: int) -> TimeSeries:
-        if not self._fit_called:
-            raise_log(Exception('fit() must be called before predict()'), logger)
-        self.predictions = []
-        for model in self.models:
-            model.fit(self.train_ts)
-            self.predictions.append(model.predict(n))
-        return np.dot(self.predictions, self.weights)
 
 
 class GROECombinationModel(CombinationModel):
